+++ src/redteam/providers/iterativeTree.ts
@@ -35,10 +35,10 @@ import { extractFirstJsonObject, safeJsonStringify } from '../../util/json';
 import { getNunjucksEngine } from '../../util/templates';
 import { sleep } from '../../util/time';
 import { shouldGenerateRemote } from '../remoteGeneration';
+import type { BaseRedteamMetadata } from '../types';
 import { PENALIZED_PHRASES } from './constants';
 import { ATTACKER_SYSTEM_PROMPT, JUDGE_SYSTEM_PROMPT, ON_TOPIC_SYSTEM_PROMPT } from './prompts';
 import { getTargetResponse, redteamProviderManager } from './shared';
-import type { BaseRedteamMetadata } from './shared';
 
 // Based on: https://arxiv.org/abs/2312.02119
